fix(agents): keep CLI-backed turns alive under stuck-session recovery
A CLI-backed agent turn that ran longer than ~6 minutes could be killed by
stuck-session recovery while it was still streaming, and its completed reply
was then never delivered to the channel.

Three producer-side gaps, all in how liveness facts reach the watchdog:

- The CLI backend counted stdout bytes for its own no-output watchdog but
  never published them as diagnostic run progress, so the freshness clock only
  advanced at tool boundaries. The core runtime already publishes per-chunk
  stream progress; the two implementations are now one shared reporter, and
  every CLI backend publishes through it.
- The CLI backend never published its resolved no-output deadline, so the
  generic 360s model-call floor preempted a backend allowance the operator had
  already configured. That deadline now rides `model.call.started` as
  `requestTimeoutMs` and bounds the generic floor, matching the direct
  provider path.
- Stale-expiry recovery ignored the `freezeAbort()` completion-wins seam that
  `supersede()` honors, so it could abort a turn whose terminal outcome was
  already committed and only delivery remained. Recovery now reports
  `keep_lane / terminal_outcome_committed` and leaves that owner to the
  existing finalization lease.

No new configuration surface: `agents.defaults.timeoutSeconds` and the
backend's `reliability.watchdog` settings now govern this path.
