+++ src/agents/cli-runner/model-call-diagnostics.ts
@@ -331,6 +331,8 @@ export function createClaudeCliModelCallDiagnostics(params: {
   prompt: string;
   systemPrompt?: string;
   transport: "paired-node-cli" | "stdio" | "stdio-live";
+  /** Backend-owned no-output deadline published to stuck-session recovery. */
+  requestTimeoutMs?: number;
   now?: () => number;
 }) {
   // Listener registration is process-stable after plugin startup. This attempt-local
@@ -473,6 +475,9 @@ export function createClaudeCliModelCallDiagnostics(params: {
         {
           type: "model.call.started",
           ...baseFields,
+          ...(params.requestTimeoutMs !== undefined
+            ? { requestTimeoutMs: params.requestTimeoutMs }
+            : {}),
         },
         privateData({ modelContent: baseModelContent() }),
       );
