+++ src/agents/embedded-agent-runner/runs.ts
@@ -12,6 +12,7 @@ import {
   abortReplyRunBySessionId,
   expireStaleReplyRunBySessionId,
   forceClearReplyOperation,
+  hasCommittedReplyOperationOutcome,
   hasReplyOperationExecutionStarted,
   isReplyRunEvidenceStaleBySessionId,
   isReplyRunActiveForSessionId,
@@ -1000,12 +1001,19 @@ export function isEmbeddedAgentRunInProgress(sessionId: string): boolean {
   return resolveEmbeddedAgentRunProgressState(sessionId) !== undefined;
 }
 
-export function resolveEmbeddedReplyActivity(
-  sessionId: string,
-): Pick<ReplyOperation, "phase" | "lastActivityAtMs"> | undefined {
+export type EmbeddedReplyActivity = Pick<ReplyOperation, "phase" | "lastActivityAtMs"> & {
+  /** Terminal outcome committed; only delivery/finalization remains. */
+  terminalOutcomeCommitted: boolean;
+};
+
+export function resolveEmbeddedReplyActivity(sessionId: string): EmbeddedReplyActivity | undefined {
   const operation = resolveActiveReplyOperationForSessionId(sessionId);
   return operation
-    ? { phase: operation.phase, lastActivityAtMs: operation.lastActivityAtMs }
+    ? {
+        phase: operation.phase,
+        lastActivityAtMs: operation.lastActivityAtMs,
+        terminalOutcomeCommitted: hasCommittedReplyOperationOutcome(operation),
+      }
     : undefined;
 }
