+++ src/auto-reply/reply/reply-run-registry.ts
@@ -46,6 +46,7 @@ export {
   waitForReplyRunSuccessorAdmission,
 } from "./reply-run-registry.registry.js";
 export {
+  hasCommittedReplyOperationOutcome,
   hasReplyOperationExecutionStarted,
   isReplyRunAbortableForSignal,
   isReplyRunEvidenceStale,
