+++ src/logging/diagnostic-session-recovery.ts
@@ -10,6 +10,7 @@ type DiagnosticSessionRecoverySkipReason =
   | "human_input_wait"
   | "runtime_owned_wait"
   | "deferred_maintenance_wait"
+  | "terminal_outcome_committed"
   | "global_lane_wait"
   | "active_lane_task"
   | "already_in_flight"
